requirements.txt
Flask==2.0.1
google-cloud-speech==2.7.0
pyaudio==0.2.11
SpeechRecognition==3.8.1
